Revert "* Clear up Level 2 warnings. Added lots of explict type casts."
This reverts commit 981bf435f3f2a4f081ba364025d6f03775f8fe34.

Conflicts:
	Trans3_2/vc/trans3/rpgcode/functions.cpp

+++ Trans3_2/vc/tkCommon/board/coords.cpp
@@ -88,8 +88,8 @@ void coords::pixelToTile(int &x, int &y, const COORD_TYPE coord, const bool bRem
 	{
 		case TILE_NORMAL:
 		{
-			x = int(floor(x / 32.0) + 1);
-			y = int(floor(y / 32.0) + 1);
+			x = floor(x / 32.0) + 1;
+			y = floor(y / 32.0) + 1;
 		} break;
 		case ISO_STACKED:
 		{
